backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Pydantic
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Azure Services
azure-cosmos>=4.9.0
azure-storage-blob>=12.24.0
azure-identity>=1.19.0
openai>=1.58.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0
cachetools>=5.5.0

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Development
black>=24.10.0
isort>=5.13.2
mypy>=1.13.0